Fix cudadrv and demo specs
