+++ CMakeLists.txt
@@ -32,7 +32,8 @@ set(AVA_GEN_GTI_SPEC OFF CACHE BOOL "Enable Gyrfalcon Plai specification")
 set(AVA_GEN_NCSDK_SPEC OFF CACHE BOOL "Enable NCSDK v1 specification")
 set(AVA_GEN_QAT_SPEC OFF CACHE BOOL "Enable QuickAssist specification")
 set(AVA_GEN_OPENCL_SPEC OFF CACHE BOOL "Enable OpenCL specification")
-set(AVA_GEN_CUDA_SPEC OFF CACHE BOOL "Enable CUDA specification")
+set(AVA_GEN_CUDADRV_SPEC OFF CACHE BOOL "Enable CUDA driver specification")
+set(AVA_GEN_CUDART_SPEC OFF CACHE BOOL "Enable CUDA runtime specification")
 set(AVA_GEN_TF_C_SPEC OFF CACHE BOOL "Enable TensorFlow (C) specification")
 set(AVA_GEN_HIP_SPEC OFF CACHE BOOL "Enable HIP specification")
 set(AVA_GEN_TEST_SPEC OFF CACHE BOOL "Enable test (libtrivial) specification")
@@ -49,7 +50,8 @@ message(STATUS "Build GTI specification:               ${AVA_GEN_GTI_SPEC}")
 message(STATUS "Build NCSDK specification:             ${AVA_GEN_NCSDK_SPEC}")
 message(STATUS "Build QuickAssist specification:       ${AVA_GEN_QAT_SPEC}")
 message(STATUS "Build OpenCL specification:            ${AVA_GEN_OPENCL_SPEC}")
-message(STATUS "Build CUDA specification:              ${AVA_GEN_CUDA_SPEC}")
+message(STATUS "Build CUDA driver specification:       ${AVA_GEN_CUDADRV_SPEC}")
+message(STATUS "Build CUDA runtime specification:      ${AVA_GEN_CUDART_SPEC}")
 message(STATUS "Build TensorFlow C specification:      ${AVA_GEN_TF_C_SPEC}")
 message(STATUS "Build HIP specification:               ${AVA_GEN_HIP_SPEC}")
 message(STATUS "Build test (libtrivial) specification: ${AVA_GEN_TEST_SPEC}")
@@ -159,7 +161,8 @@ ExternalProject_Add(ava-spec
         -DAVA_GEN_NCSDK_SPEC:BOOL=${AVA_GEN_NCSDK_SPEC}
         -DAVA_GEN_QAT_SPEC:BOOL=${AVA_GEN_QAT_SPEC}
         -DAVA_GEN_OPENCL_SPEC:BOOL=${AVA_GEN_OPENCL_SPEC}
-        -DAVA_GEN_CUDA_SPEC:BOOL=${AVA_GEN_CUDA_SPEC}
+        -DAVA_GEN_CUDADRV_SPEC:BOOL=${AVA_GEN_CUDADRV_SPEC}
+        -DAVA_GEN_CUDART_SPEC:BOOL=${AVA_GEN_CUDART_SPEC}
         -DAVA_GEN_TF_C_SPEC:BOOL=${AVA_GEN_TF_C_SPEC}
         -DAVA_GEN_HIP_SPEC:BOOL=${AVA_GEN_HIP_SPEC}
         -DAVA_GEN_TEST_SPEC:BOOL=${AVA_GEN_TEST_SPEC}
